extract option to separate file
